small kotlin refactoring
