--- resultcapture-service/src/main/kotlin/com/devotion/capture/ValidationException.kt
@@ -1,7 +1,0 @@
-package com.devotion.capture
-
-class ValidationException(message: String) : RuntimeException(message) {
-    companion object {
-        private val serialVersionUID = -3685317928211708951L
-    }
-}
