--- resultcapture-service/src/main/kotlin/com/devotion/capture/service/UserService.kt
@@ -1,11 +1,0 @@
-package com.devotion.capture.service
-
-import com.devotion.capture.model.AnonimousUser
-
-import javax.servlet.http.HttpServletRequest
-
-interface UserService {
-
-    fun resolveUser(request: HttpServletRequest): AnonimousUser
-
-}
