+++ resultcapture-service/src/main/kotlin/com/devotion/capture/service/repository.kt
@@ -1,14 +1,16 @@
 package com.devotion.capture.service
 
 import com.devotion.capture.model.QuestionAnswer
+import com.devotion.capture.model.SurveyResult
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface QuestionAnswerRepository : MongoRepository<QuestionAnswer, String> {
-
-    fun findBySurveyIdAndQuestionId(surveyId: String, questionId: Int?): List<QuestionAnswer>
-
+interface SurveyResultRepository : MongoRepository<SurveyResult, String> {
+    fun findAnswersBySurveyId(surveyId: String): List<QuestionAnswer>
 }
 
-
+@Repository
+interface QuestionAnswerRepository : MongoRepository<QuestionAnswer, String> {
+    fun findBySurveyIdAndQuestionId(surveyId: String, questionId: Int?): List<QuestionAnswer>
+}
